Support `atmos terraform apply --from-plan` with additional flags
Right now, when atmos generates the plan-file arg for `terraform apply`,
it puts it first, and then appends any additional args or flags that may
be specified by the user. Terraform is very particular about the order
of args and flags, and crashes with an error if flags follow the
plan-file args. Instead, put the plan-file arg at the end of the
command, after any additional flags or args.

For example, with this change `atmos terraform apply -s $STACK
$COMPONENT planfile -- -parallelism=1` will call `terraform apply
-parallelism=1 planfile` instead of `terraform apply planfile
-parallelism=1`.