Always template vendor source and targets (#712)
This change improves the templating within vendor manifests slightly: It
officially adds support for the `Component` field to both `source` and
`targets`.

These features were already supported but mostly undocumented and hidden
behind an implicit switch: The templating was only triggered if the `Version`
field was set. Which was also the only officially supported field.

In reality though all fields from the current source definition were
available but in the state they were currently in, depending on the
order of the templates.

With this change
 * It is clearly documented which fields are supported in which YAML
   values.
 * Only the two static fields are supported.
 * The values are always templated.

Theoretically this could be a breaking change if somebody used no
`version` field but curly braces in their paths. Or relied on the
half-populated source data structure to refer to unsupported fields.
If xkcd 1172 applies it should be possible to amend this logic to add
more officially supported fields.

Co-authored-by: Andriy Knysh <aknysh@users.noreply.github.com>