fix: remove redundant MergeInConfig call that broke Windows tests
The tempViper.MergeInConfig() call was attempting to merge the config into itself,
which was causing YAML template functions to return null values on Windows.

This bug was introduced in PR #1447 and caused test failures in:
- TestYamlFuncTerraformOutput
- TestProcessCustomYamlTags

The fix removes this unnecessary call since we're already processing the config
correctly by marshaling tempViper to YAML and then merging it into the main
Viper instance.